+++ config/routes.rb
@@ -1,4 +1,4 @@
 # Plugin's routes
 # See: http://guides.rubyonrails.org/routing.html
 
-resources :weekly_schedules, except: [:index]
+resources :recurring_tasks, except: [:index]