+++ init.rb
@@ -11,7 +11,7 @@ Redmine::Plugin.register :redmine_recurring_tasks do
   name 'Redmine Recurring Tasks'
   author 'Southbridge'
   description 'Plugin for creating scheduled tasks from template'
-  version '0.0.2'
+  version '0.1.0-dev'
   url 'https://github.com/centosadmin/redmine_recurring_tasks'
   author_url 'https://github.com/centosadmin'
 
@@ -25,7 +25,7 @@ Redmine::Plugin.register :redmine_recurring_tasks do
   )
 
   project_module :redmine_recurring_tasks do
-    permission :view_schedule,   weekly_schedules: :show, read: true
-    permission :manage_schedule, weekly_schedules: [:new, :edit, :destroy, :update], require: :loggedin
+    permission :view_schedule,   recurring_tasks: :show, read: true
+    permission :manage_schedule, recurring_tasks: [:new, :edit, :destroy, :update], require: :loggedin
   end
 end