+++ lib/redmine_recurring_tasks/issue_checker.rb
@@ -9,18 +9,9 @@ module RedmineRecurringTasks
       @settings = settings
     end
 
-    # @return [Array<WeeklySchedule>] a schedules which should be executed
+    # @return [Array<RecurringTask>] a schedules which should be executed
     def schedules
-      date = Time.now
-      week_day = date.strftime('%A').downcase
-
-      WeeklySchedule.where("#{week_day} = true").map do |schedule|
-        time_came = schedule.time.strftime('%H%M%S') <= date.strftime('%H%M%S')
-
-        if time_came && (schedule.last_try_at.nil? || schedule.last_try_at.strftime('%Y%m%d') < date.strftime('%Y%m%d'))
-          schedule
-        end
-      end.compact
+      RecurringTask.schedules
     end
 
     # @return [void]